Massive refactor to make Stimulus code more understandable
The previous version was very confusing as it had four controllers, and the controllers had mixed responsibilities.  This led to it being difficult to know where to go to apply a change.

The new code tries to fix this by having only two controllers:
* expense-entry.  This deals with any changes that happen within an expense entry line.  Principally, this deals with changing the category, unit cost, or quantity. 
* expense-claim.  This controller deals with all changes that affect the overall expense claim, even if the change originates in an expense row.  So examples include updating the overall claim cost.  This controller keeps track of which expense entry is being changed, and once we move to a new expense entry row, submits any changed rows.

This approach has allowed us to simply also the structure of the HTML on the page.

I have been more consistent in naming.  Everything is either an expense-entry or an expense-claim.  Interchangeable references between expense-entry and expense-row have gone.

Moved to using HAML and have also setup RSpec, although the actual tests are empty right now.